% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_function.R
\name{factorial_function}
\alias{factorial_function}
\title{A Factorial Function}
\usage{
factorial_function(n)
}
\description{
This function does a factorial given a few conditions
}
\examples{
factorial_function()
}
\keyword{factorial}
